requirements.txt
anyio==4.10.0
asgiref==3.9.1
black==25.1.0
click==8.2.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
flake8==7.3.0
gunicorn==22.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy==1.17.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
